Addressing issue no #198
Have separated isTouchDevice form DeveloperSignup.js
